DroneProcess...

+++ Assets/_myGame/_scripts/Weapon/Concrete/WeaponScriptableObject.cs
@@ -5,15 +5,13 @@ using UnityEngine;
 
 [CreateAssetMenu(fileName = "CreateWeapon", menuName = "ScriptableObjects/Weapons", order = 1)]
 public class WeaponScriptableObject : ScriptableObject
-{
-    
+{    
    public float attackDelay;
    public float attackDamage;
-   public float maxTargetCount;
+   public int maxTargetCount;
    public float range;
    public WeaponAnimationController weaponAnim;
    public GunType GunType;
-
 }
 
 public enum GunType
